Improved dependency management
By harnessing the power of gradle and maven, dependencies can now be managed much more easily and in a much more elegant manner. Glmath and redstoneflux source files were removed. Both glmath and redstone flux are now declared as dependencies in build.gradle. Redstone Flux is required as an additional mod to be present in the mod environment (the mod won't load without it), and glmath is packed in the jar after compilation. The final effect is a little bit faster build time, since all the libraries come pre-compiled.